#10 Save All for LabVIEW 2020
Saved all for LabVIEW 2020 and marked all the VIs to separate source code from compiled code

Also, disabled the code inside the ErrorLogDialog.vi because those VIs are missing. I left a note with the following: "#CodeReview-To-Do [Fabiola De la Cueva - Feb-22-2025] This VI is missing dependencies and it does not seem to be called by the rest of the code.
I am disabling it so I can save the rest of the VIs for LabVIEW 2020 and already reached out to Steve to see what he
wants me to do with this VI." and I created #11 issue for Steve to address this